requirements.txt
appnope==0.1.0
arrow==0.13.1
astroid==2.2.5
attrs==19.1.0
backcall==0.1.0
binaryornot==0.4.4
bleach==3.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cookiecutter==1.6.0
coverage==4.5.3
cycler==0.10.0
decorator==4.4.0
defusedxml==0.5.0
Django==2.2
django-cors-headers==2.5.3
django-rest-framework==0.1.0
djangorestframework==3.9.3
entrypoints==0.3
future==0.17.1
idna==2.8
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.15
jedi==0.13.3
Jinja2==2.10
jinja2-time==0.2.0
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
Markdown==3.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
mistune==0.8.4
music21==5.5.0
nbconvert==5.4.1
nbformat==4.4.0
notebook==5.7.6
numpy==1.16.2
pandas==0.24.2
pandocfilters==1.4.2
parso==0.3.4
pbr==5.2.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==6.0.0
poyo==0.4.2
prometheus-client==0.6.0
prompt-toolkit==2.0.9
psycopg2==2.8.2
ptyprocess==0.6.0
Pygments==2.3.1
pylint==2.3.1
pyparsing==2.3.1
pyrsistent==0.14.11
python-dateutil==2.8.0
pytz==2019.1
pyzmq==18.0.1
qtconsole==4.4.3
requests==2.21.0
scipy==1.2.1
Send2Trash==1.5.0
six==1.12.0
sqlparse==0.3.0
stevedore==1.30.1
terminado==0.8.1
testpath==0.4.2
tornado==6.0.2
traitlets==4.3.2
typed-ast==1.3.1
urllib3==1.24.3
virtualenv==16.4.3
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
wcwidth==0.1.7
webencodings==0.5.1
whichcraft==0.5.2
widgetsnbextension==3.4.2
wrapt==1.11.1
